add layer service and string extension for splittakelast
