+++ uic-addin.csproj
@@ -141,7 +141,9 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="Attributes\FipsCodeAttribute.cs" />
+    <Compile Include="Extensions\StringExtensions.cs" />
     <Compile Include="Models\FacilityModel.cs" />
+    <Compile Include="Services\LayerService.cs" />
     <Compile Include="Services\QueryService.cs" />
     <Compile Include="Services\ThreadService.cs" />
     <Compile Include="ViewModels\PropertyChangedWithValidation.cs" />
